refactor: reorganize tests into declarative e2e and programmatic integ
- Move testscript scenarios from internal/integ/ to tests/e2e/
- Move programmatic tests from internal/integ/ to tests/integ/
- Rename unit/integration terminology to local/api for clarity
- Update build tags from 'integration' to 'api'
- Rename test functions: TestScriptsUnit → TestScriptsLocal, TestScriptsIntegration → TestScriptsAPI
- Update GitHub Actions workflow for new structure and build tags
- Update documentation to reflect new organization

This provides cleaner separation between:
- tests/e2e/: Declarative testscript scenarios (local vs api)
- tests/integ/: Programmatic Go tests with suite helpers

Both test approaches are end-to-end from CLI user perspective, just different implementation styles.
