+++ tests/e2e/scenarios/local/basic_no_api.txtar
