+++ tests/e2e/scenarios/local/config_isolated.txtar
