+++ tests/integ/go.mod
@@ -1,0 +1,16 @@
+module github.com/exoscale/cli/internal/integ
+
+go 1.23
+
+require (
+	github.com/rogpeppe/go-internal v1.14.1
+	github.com/stretchr/testify v1.9.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/sys v0.26.0 // indirect
+	golang.org/x/tools v0.26.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
